Add status for published version of feeds ### Checklist
- [ ] Appropriate branch selected _(all PRs must first be merged to `dev` before they can be merged to `master`)_
- [ ] Any modified or new methods or classes have helpful JSDoc and code is thoroughly commented
- [ ] The description lists all applicable issues this PR seeks to resolve
- [ ] The description lists any configuration setting(s) that differ from the default settings
- [ ] All tests and CI builds passing
- [ ] The description lists all relevant PRs included in this release _(remove this if not merging to master)_
- [ ] e2e tests are all passing _(remove this if not merging to master)_

### Description

Adds a new column to the feed source table to show the status of a published feed.

Please explain the changes you made here and, if not immediately obvious from the code, how they resolve any referenced issues (and PRs if merging to master). Be sure to include all issues being resolved and any special configuration settings that are need for the software to run properly with these changes.